UI: Simplify `ClementineApp` removing overcomplicated struct for widget
